refactor: upload qr modal (#437)
